Add Apache 2.0 license, README, and workflows
This commit adds an Apache 2.0 license file and a README file to the repository. It also includes the setup of several Github workflows for automating tasks such as Python style checking, internal ROS 2 package building & testing, synchronization of the README to a Notion page, and the use of pre-commit checks.