Update git installation method in pre-commit workflow
This commit modifies the method used for git installation in our pre-commit workflow. Instead of simply installing git using apt-get, it now forces the installation of the latest version of git through the git-core/ppa repository. This helps ensure that our workflow is always using the most current, stable features of git.